Do not index using specific NE version numbers in topojson
Otherwise, we have to update the UI code every time we use a new
NE topojson. The API will deliver the intended version so there
is no reason to hard-code that information in the UI as well.

NOTE: DO NOT MERGE TO MAIN UNTIL AFTER THE 5.1 TOPOJSON IS
DEPLOYED ON THE PROD API CONTAINER.